Fixed signal handling and updated xcode settings

+++ Tests/TestObject.h
@@ -26,7 +26,7 @@ NSArray* TestObjectCalledSelectors(TestObject *object);
 #pragma mark - Void Return Calls
 
 - (void)voidMethodCallWithoutParameters;
-- (void)voidMethodCallWithIntParam1:(int)i1 intParam2:(int)i2;
+- (void)voidMethodCallWithIntParam1:(NSInteger)i1 intParam2:(NSInteger)i2;
 - (void)voidMethodCallWithDoubleParam1:(double)d1 doubleParam2:(double)d2;
 - (void)voidMethodCallWithObjectParam1:(id)o1 objectParam2:(id)o2;
 - (void)voidMethodCallWithSelectorParam1:(SEL)s1 selectorParam2:(SEL)s2;
